CAMEL-23850: address further review feedback on route-diagram/topology source dump
- Extract the duplicated transient-boot source-dump logic from
  CamelRouteDiagramAction and CamelRouteTopologyAction into a shared
  ActionBaseCommand#dumpRoutesFromSource helper (returns workDir + boot
  start instant via a SourceDump record).
- Drop the misleading volatile qualifier on the single-threaded pid field
  in both actions.
- Replace the ad-hoc "field never used" note on the picocli completion-only
  filePaths field with @SuppressWarnings("unused") to silence static
  analysis.
- Fix the Thread.sleep migration in CamelRouteDiagramActionMultiFileTest
  (handle InterruptedException, drop now-unused imports) and comment that
  the sleeps simulate delayed dump-file writes.
